server/requirements.txt
Flask==3.1.1
matplotlib==3.7.2
numpy==1.24.4
Pillow==11.2.1
Requests==2.32.3
torch==2.4.1
torchvision==0.20.0
tqdm==4.67.1
Werkzeug==3.1.3
